backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Authentication
python-social-auth[django]>=0.1.0
social-auth-app-django>=5.4.0
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0

# WebSockets / Real-time
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Database
psycopg2-binary>=2.9.9
django-redis>=5.4.0

# Google Sheets Integration
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
gspread>=5.12.0

# Steam API
python-steam>=1.0.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2
pytz>=2023.3
celery>=5.3.0
redis>=5.0.0

# Server Status (Game Server Query)
python-a2s>=1.3.0

# Development
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0
pytest-django>=4.5.0
